Extract WasabiApplication
